fix warning
svn path=/trunk/playground/devtools/icemon-kde3/; revision=559972

+++ src/joblistview.cc
@@ -329,7 +329,7 @@ void JobListView::slotExpireFinishedJobs()
     FinishedJobs::iterator it = mFinishedJobs.begin();
     for ( const FinishedJobs::iterator itEnd = mFinishedJobs.end(); it != itEnd; ++it )
     {
-        if ( currentTime - ( *it ).first < mExpireDuration )
+        if ( currentTime - ( *it ).first < (uint)mExpireDuration )
             break;
 
         removeItem( ( *it ).second );
